# vsr: adaptive repair timeouts