### Summary

This PR implements adaptive repair timeouts (ARTs), to make repair robust in the face of changing network topology and latencies. 

A key deficiency in WAL repair currently is that it _randomly chooses a remote replica_ to request a prepare from. This doesn't work well with one of our most common repair workloads -- when a lagging replica transitions to a new checkpoint, it must repair the portion of the WAL that its missing. If we request a prepare from another lagging replica (3/6 replicas tend to lag behind the fastest quorum, even with adaptive replication), we have to wait a repair timeout to re-request it, as we only request a prepare once between two timeouts. With our random replica selection, we may even go multiple repair timeouts without the prepare. Requesting a prepare only once between repair timeouts also meant that a replica's repair budget is well utilized when it has many prepares to repair, but mostly unutilized if it only had very few missing prepares.

Adaptive repair timeouts improve upon this, since we now rely on measured repair latency to choose a replica to repair from. Specifically, replicas attempt to direct their request to a replica that has previously provided us with a good repair quality of service (low repair latency). 

Additionally, this PR makes a crucial change to how replicas utilize their repair budgets. Now, replicas _always_ fully utilize their repair budget. Even, if a single prepare is missing, it is requested from _all_ remote replicas (with at most one request to a remote replica). The rationale here is as follows:

* By speeding up the repair for the single missing prepare (by sending redundant requests), even though we use some extra disk/network bandwidth on remote replicas, it _saves future bandwidth_ since it reduces the likelihood of the replica falling further behind. 
* Since we expect the cluster to handle the repair load when a replica is lagging by more and completely saturates the budget, it would be reasonable to expect that even if there’s a single missing prepare. And perhaps even better, since we predictably keep the repair bandwidth utilisation similar despite how much we’re lagging.


### Implementation Details

We implement adaptive repair timeouts without explicitly maintaining a timeout with dynamically varying interval!Instead, we maintain the instant at which a prepare is requested for repair, and which replica it has been requested from. We use this instant to compute the repair latency for each remote replica, by measuring the duration up till either:
* The requested prepare is received from the remote replica. Or,
* The requested prepare hits its _expiry duration_.
  * Expiry is enforced by `journal_repair_budget_timeout`, which fires every 10ms, peeks into each budget, and expires requests that are due for expiration.  At expiry, we penalize the replica to which the request expired, by adding the duration that we waited  for, to its current repair latency. 
  * The *expiry duration* for a request is a multiple of the repair latency to that replica (set to 2x in the PR), *and* is also capped to 500ms, to avoid the scenario where a partitioned replica with missing prepares triggers repair over and over on timeout, and unboundedly increases the repair latency for all replicas.
 
The criterion for selecting a replica to repair from is simple, we choose the replica with the *lowest repair latency, that also has available budget* (we allocate a budget of 2 for each remote replica, so as to not overwhelm a remote replica). Additionally, a fraction of ops are experimental ops, for which we try an alternative repair routes. This avoids potential resonance wherein we keep requesting from a permanently crashed replica with the best repair latency. This is because we don't penalize the repair latency once it exceeds the maximum expiry duration, so if a crashed replica has the best latency, it may remain that way forever.

### Deterministic Performance Testing (Median over 100 runs)

#### 30% packet loss

| Metric               | main | ART | Δ (absolute) | Δ (%)      |
|----------------------|-----------|------------|--------------:|-----------:|
|Ticks | 294613   | 280103    |-14510     | **-4.92%** |

Packet loss is a good way to simulate a scenario where our old replica selection criterion would send a RequestPrepare to a replica that doesn't have the prepare. In this case, the requesting replica's RequestPrepare/the remote replica's Prepare may be dropped.

**With ARTs, get a ~5% end-to-end performance boost in the presence of 30% packet loss, because we now _hedge_ RequestPrepares for an op, as long as we have the budget to do so.** Where earlier, a dropped RequestPrepare/Prepare could lead to at most a 100ms delay before we send _another_ RequestPrepare. Now, we send more RequestPrepare requests as long as we have the budget!

#### Missing Replicas


| Metric               | Missing Replicas| main | ART | Δ (absolute) | Δ (%)      |
|----------------------|-|-----------|------------|--------------:|-----------:|
|Ticks |1| 103494   | 101863    | -1631    | **-1.57%** |
|Ticks |2| 116372   | 112782    | -3590    | **-3.08%** |

My initial intuition was that with ART, we should see predictable performance regardless _regardless of the number of crashed replicas_. I since realized that is not possible, since missing replicas _also_ reduce our repair throughput, we have fewer replicas to repair from! 

However, ART _does_ increase end-to-end performance by **~3%** since the replicas that are missing are automatically deprioritized, whereas the algorithm on main randomly selects replicas for repair, so it sends relatively more requests to the missing replicas. With R2 and R3 as missing replicas, here is the count of RequestPrepares sent by R4 (randomly chosen):

| Remote Replica               | Status| ART | main |
|----------------------|-|-----------|------------|
|0 |🌱 | 2648   | 2160    | 
|1 👑 |🌱 | 4085   | 2180    | 
|2 |💀| 608   | 2179    | 
|3 |💀| 621   | 2160    | 
|5 |🌱| 2801   | 2140    | 

We can see that with ART, the maximum # of repair requests were directed towards the primary, which gave the best quality of service to R4 since it is guaranteed to have all prepares. On the other hand, with ART, **~5.6%** of requests were directed towards the crashed replicas (R2 and R3), while with main, **~20%** of requests were directed towards them (as it randomly selects replicas).