@chaitanyabhandari 2025-10-03
One final set of measurements on a real cluster of 6 replicas, running `./tigerbeetle benchmark --query-count=0 --transfer-count=10_000_000 --clients=8 --transfer-batch-size=2`. Nothing _too_ exciting there in terms of end-to-end performance, except for a drop in p100 from `1142 ms → 736 ms`, for which I'm planning on taking some more measurements to validate if it is just an outlier (the benchmark took 8 hours so it will take a while to validate):

```C
transfer batch size = 2 txs
transfer batch delay = 0 us
load accepted = 486 tx/s → 486 tx/s
batch latency p1   = 4 ms → 4 ms
batch latency p50  = 30 ms → 30 ms
batch latency p99  = 121 ms → 123 ms
batch latency p100 = 1142 ms → 736 ms
```

Also measured the # of ignored prepares when a replica is lagging, which is the main contributor for repair traffic as the lagging replica then has to repair these prepares when it transitions to the new checkpoint. We see a reduction from `5432` → `4178` for the replica 4, which presumably is lagging the most: 


| Replica               | Metric| ART | main |
|----------------------|-|-----------|------------|
|0 |# of ignored prepares | 32   | 18    | 
|1 |# of ignored prepares | 0   | 0    | 
|2 |# of ignored prepares| 112   | 67    | 
|3 |# of ignored prepares| 1156   | 1110    | 
|4 |# of ignored prepares| 4178   | 5432    | 
|5 |# of ignored prepares| 116   | 32    |